refactor: profile folder
