+++ src/components/pages/profile/player/basicInfo/BasicInfo.css
