+++ src/pages/edit/Edit.css
@@ -1,0 +1,29 @@
+.edit h1 {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 1.5em;
+    margin-top: 1em;
+}
+
+.edit h1 span{
+    font-size: 0.75em;
+    color: var(--text-low);
+    margin-left: 0.4em;
+    margin-top: 0.25em;
+}
+
+.edit .content {
+    display: flex;
+    gap: 4em;
+    padding: 2em 4em;
+}
+
+.edit .content .column-1,
+.edit .content .column-3 {
+    flex: 1;
+}
+
+.edit .content .column-2 {
+    flex: 2;
+}
