+++ src/pages/edit/Edit.jsx
@@ -1,0 +1,27 @@
+import React from 'react'
+import Footer from '../../components/layout/footer/Footer'
+import './Edit.css'
+
+const Edit = () => {
+    return (
+        <div className='edit'>
+            <h1>Edit hattrick database</h1>
+            <section className='content'>
+                <section className='column-1'>
+                    <h2>Teste</h2>
+                </section>
+                
+                <section className='column-2'>
+
+                </section>
+                
+                <section className='column-3'>
+
+                </section>
+            </section>
+            <Footer />
+        </div>
+    )
+}
+
+export default Edit
